pyproject.toml
[project]
name = "cricinfo"
version = "0.1.0"
description = ""
authors = [
    {name = "Amogh Shukla",email = "amoghshukla548@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.136.0,<0.137.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "bcrypt (==4.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
